requirements.txt
pandas
openpyxl
pyodbc
gspread
google-auth
playwright
pywin32